Implement `LlmDialogController`: fully finish `converse` and `converseCold` methods (#6)
* Create sendMessages() method in LlmProvider

* Install & configure vitest

* Create Questionnaire class and its DTO

* Implement draft version of FewShotPromptingStrategy

The draft implementation serves as a basis for structured interface-based family of classes.

* Create a test for FewShotPromptingStrategy

* Create custom application error

* Add JSON files with personas, intructions, and prompt templates

* Implement `PromptTemplate` class

This class helps set variables present in the prompt template.

* Use prompt templates from JSON files in `FewShotPromptingStrategy`

* Allow `"resolveJsonModule"` in tsconfig.json

* Add comments & fixes to `PromptTemplate` and update prompt template definitions.

* Create Wendy persona and 2 instruction prompt templates

* Update `FewShotPromptingStrategy` to use `PromptTemplate`

* Add outputFormat to buildExampleConversationInstruction

* Refactor `LlmDialog`, `LlmProvider`, and `OpenAILlmProvider`

* Add comment to `PromptTemplate`

* Create `ChatMessage` & `llmRegex`

* Create DB data models & update their imports

* Make data models' constructors accept objects

* Move all logical components into 'src/app' folder

* Add dto for `RecurringChatNotification`

* Start developing `IRepository` APIs

* Install `openai-gpt-token-counter` via npm

This library allows to count the number of tokens based on context messages.

* Add cause argument to `ApplicationError`

* Install `tiktoken` library for token counting

* Add `countTokens` method to `LlmProvider`

* Add method for counting tokens for text and messages

Sometimes we want to be able to count the tokens number either for
messages or text, thus 2 implementations required.

* Make `Persona` a base class for direct personas implementor

It is useful to allow persona parameterization.

* Add `dialogInitializationInstructionPromptTemplate` prompt template & add additional prompt variables

Added `{{SUMMARIZED_MESSAGES}}` and `{{CONVERSATION_MESSAGES}}` prompt variables.

* Make `ChatMessage` and `Questionnaire` implement `IPromptifiable`

It is useful to have a quick way of crafting formatted prompt-ready strings from these DTO classes.

* Remove import of `IByUserIdSearchableRepository` from `UserRepository`

* Refactor `PromptTemplate` and rename `CHAT_MESSAGE` to `LAST_CHAT_MESSAGE`

* Implement role-sealed derived classes of `LlmChatMessage`

There are 3 classes that client must use to work with llm chat messages:
1. `UserLlmChatMessage`: has `User` role.
2. `AssistantLlmChatMessage`: has `Assistant` role.
3. `SystemLlmChatMessage`: has `System` role.

* Rename `LlmDialog` to `LlmDialogManager` and implement `converse` method

For now, the `LlmDialogManager` class requires caller to keep the chat history under token limit themselves.
I plan to implement a class that will be responsible for history truncation.

* Create `IPromptifiable` (ATTENTION: it must have been added in previous commit!)

* Implement utility test to check promptified `ChatMessage`

* Update DB Repositories

Add additional interface into `Repositories.ts`,
and implement a method `getPaginatedSortedMessages`
in `ChatMessageRepository` for paginated messages retrieval.

* Move `LlmChatHistory` into different file & add `createGeneralDialogInstructionPrompt` method to `LlmDialogManager`

In the future, the `LlmDialogManager` should be refactored: methods that
work with prompt templates should extract to a different class.

* Create a drafted version of `LlmDialogController`

`LlmDialogController` is responsible for LLM chat history creation and
subsequent LLM prompting.

* Rename `ChatMessage` to `ChatMessageModel`

Before, the name clashed with the Prisma's `ChatMessage` DB class.
Same goes for all the rest models. Renaming for the rest classes comes shortly.

* Rename `Questionnaire` to `QuestionnaireModel`

* Rename `RecurringChatNotification` to `RecurringChatNotificationModel`

* Rename `User` to `UserModel`

* Create `ChatMessageModel`'s ctor of `ChatMessage`

* Create `QuestionnaireModel`'s ctor of `Questionnaire`

* Move `userId` in `ChatMessageModel` to its DTO

* Move `userId` in `QuestionnaireModel` to its DTO

* Make `RecurringChatNotificationModel`'s ctor accept `RecurringChatNotification`

* Make `UserModel`'s ctor accept `User`

* Create `IByUserIdManyRetrievableRepository` & `IByUserIdUniqueRetrievableRepository` interfaces

There must be a distinction between searching MANY by user id and searching UNIQUE by user id.

* Add comment about `LlmChatHistory`'s usage API

* Create `QuestionnaireRepository`

It is responsible for CRUD operation over database table with users' questionnaires.

* Store advancements in `LlmDialogController` implementation

There is a basic UNTESTED impl of `LlmDialogController`.
The next step is to test whether it works as intended.

* Move `prompting` directory into `llm` directory

* Remove `@unique` from `ChatMessage` model in schema.prisma

It effectively made multiple messages per user impossible. Definitely not what we want.
Likely, a typo from the initial definition.

* Remove unused imports from `llm.test.ts`

* Stash out changes into llm.test.ts & resolve merge conflicts

* Create build script in npm

* Compile only sources from src folder excluding test

* Create `CollectionsUtils.ts` file with util functions over collection-like data structures

* fix(PromptTemplate): fix variable replacement bug

Bug: deleted only first entry of a variable.
Fix: replace all entries of a variable with its value via JS RegExp.

* fix(repositories): rename method name to `getByUserId`

Issue: the name was `getUserById`, yet it is not the idea of a method.
Fix: rename the method to `getByUserId`.

* Change formatting in `promptify` realizations of DB models

Remove extra states at the beginning of new lines.

* fix(LlmDialogManager): insert system prompt as 1st in the history

Bug: forgot to insert the initial system prompt into the history that is sent to LLM.
Fix: insert it as the first prompt in the history.

* Remove timeout on test of `LlmDialogController` & add enhancements in class' implementation

* Provide LLm's persona as argument to `LlmDialogController`

* fix(PromptTemplates): textual incorrectness

Issue: used "her" to refer the summary.
Fix: paraphrase and replace with "them".

* Refactor implementation of `LlmDialogController`

Made first steps towards refactoring.

* Comment out tests that test private API of `PromptTemplate`

Commented tests work but the methods tested are private (i.e., it runs but IDE complains).

* Refactor `LlmDialogController` to use Policy pattern to isolate algorithmic behaviours

Created `IHistoryConstructionPolicy` & `IMessagesSplitPolicy` that define how to construction
a history and how to split the messages between summarized and plain ones respectively.

* Start implementation of `startColdConversation` method of `LlmDialogController`

* Return `AssistantLlmChatMessage` from converse methods of `LlmDialogController`

* Create instructions for message summarizing

* Implement message summarization in `LlmDialogManager`

* Use message summarization & add comments in `LlmDialogController`

* fix(TwoMessageHistoryConstructionPolicy): use descending pagination of messages

Issue: wrong sorting by message sent date: ascending instead of descending.
Fix: change to descending.

* Make minor comment changes

* Add `env.ts` into `tsconfig.json` file & change root dir