feat(bot): Implement bot API
Now, user can interact with LLM via Telegram chat. The only requirement is to have a questionnaire in the database.

Note: user's questionnaire yet cannot be created with Telegram WebApp, as it required a public hosting to send data to the bot.
Questionnaire is created manually in the database for testing.